New feature: enable setting lock dir permissions
The directory that is created by the acquire function (i.e. os.mkdir command) uses the default setting of mode=0o777 (i.e. world writable) with no way to change it. It is a security best practice to not create world writable objects, and vulnerability scanners are known to check for world writable objects. A mode parameter with a default value of 0o777 for backwards compatibility was added to the class constructor, and the mode parameter was added to the os.mkdir command to allow the mode value to be customized.